+++ src/Ext/WarheadType/Detonate.cpp
@@ -104,7 +104,7 @@ void WarheadTypeExt::ExtData::Detonate(TechnoClass* pOwner, HouseClass* pHouse, 
 
 	this->Crit_CurrentChance = this->GetCritChance(pOwner);
 
-	if (this->PossibleCellSpreadDetonate || this->Crit_CurrentChance > 0.0)
+	if (this->PossibleCellSpreadDetonate || (this->Crit_CurrentChance.size() == 1 && this->Crit_CurrentChance[0] > 0.0) || this->Crit_CurrentChance.size() > 1)
 	{
 		this->Crit_Active = false;
 
@@ -145,7 +145,7 @@ void WarheadTypeExt::ExtData::DetonateOnOneUnit(HouseClass* pHouse, TechnoClass*
 	if (this->RemoveMindControl)
 		this->ApplyRemoveMindControl(pHouse, pTarget);
 
-	if (this->Crit_CurrentChance > 0.0 && (!this->Crit_SuppressWhenIntercepted || !bulletWasIntercepted))
+	if (this->Crit_CurrentChance.size() > 0 && (!this->Crit_SuppressWhenIntercepted || !bulletWasIntercepted))
 		this->ApplyCrit(pHouse, pTarget, pOwner, pTargetExt);
 
 	if (this->Convert_Pairs.size() > 0)
@@ -269,16 +269,6 @@ void WarheadTypeExt::ExtData::ApplyRemoveDisguiseToInf(HouseClass* pHouse, Techn
 
 void WarheadTypeExt::ExtData::ApplyCrit(HouseClass* pHouse, TechnoClass* pTarget, TechnoClass* pOwner, TechnoExt::ExtData* pTargetExt = nullptr)
 {
-	double dice;
-
-	if (this->Crit_ApplyChancePerTarget)
-		dice = ScenarioClass::Instance->Random.RandomDouble();
-	else
-		dice = this->Crit_RandomBuffer;
-
-	if (this->Crit_CurrentChance < dice)
-		return;
-
 	if (!pTargetExt)
 		pTargetExt = TechnoExt::ExtMap.Find(pTarget);
 
@@ -293,8 +283,36 @@ void WarheadTypeExt::ExtData::ApplyCrit(HouseClass* pHouse, TechnoClass* pTarget
 		if (pSld && pSld->IsActive() && pSld->GetType()->ImmuneToCrit)
 			return;
 
-		if (pTarget->GetHealthPercentage() > this->Crit_AffectBelowPercent)
+		if (this->Crit_AffectBelowPercent.size() > 0 && pTarget->GetHealthPercentage() > this->Crit_AffectBelowPercent[0])
 			return;
+	}
+
+	unsigned int level = 0;
+
+	if (this->Crit_AffectBelowPercent.size() > 0)
+	{
+		for (; level < this->Crit_AffectBelowPercent.size() - 1; level ++)
+		{
+			if (pTarget->GetHealthPercentage() > this->Crit_AffectBelowPercent[level + 1])
+				break;
+		}
+	}
+
+	double dice;
+
+	if (this->Crit_ApplyChancePerTarget)
+		dice = ScenarioClass::Instance->Random.RandomDouble();
+	else
+		dice = this->Crit_RandomBuffer;
+
+	if (this->Crit_CurrentChance.size() == 1)
+	{
+		if (this->Crit_CurrentChance[0] < dice)
+			return;
+	}
+	else if (this->Crit_CurrentChance.size() <= level || this->Crit_CurrentChance[level] < dice)
+	{
+		return;
 	}
 
 	if (pHouse && !EnumFunctions::CanTargetHouse(this->Crit_AffectsHouses, pHouse, pTarget->Owner))
@@ -316,7 +334,12 @@ void WarheadTypeExt::ExtData::ApplyCrit(HouseClass* pHouse, TechnoClass* pTarget
 		GameCreate<AnimClass>(this->Crit_AnimList[idx], pTarget->Location);
 	}
 
-	auto damage = this->Crit_ExtraDamage.Get();
+	int damage = 0;
+
+	if (this->Crit_ExtraDamage.size() == 1)
+		damage = Crit_ExtraDamage[0];
+	else if (this->Crit_ExtraDamage.size() > level)
+		damage = Crit_ExtraDamage[level];
 
 	if (this->Crit_Warhead.isset())
 		WarheadTypeExt::DetonateAt(this->Crit_Warhead.Get(), pTarget, pOwner, damage);
@@ -430,12 +453,15 @@ void WarheadTypeExt::ExtData::ApplyAttachEffects(TechnoClass* pTarget, HouseClas
 	AttachEffectClass::DetachByGroups(this->AttachEffect_RemoveGroups, pTarget, this->AttachEffect_CumulativeRemoveMinCounts, this->AttachEffect_CumulativeRemoveMaxCounts);
 }
 
-double WarheadTypeExt::ExtData::GetCritChance(TechnoClass* pFirer) const
+std::vector<double> WarheadTypeExt::ExtData::GetCritChance(TechnoClass* pFirer) const
 {
-	double critChance = this->Crit_Chance;
+	std::vector<double> critChance = this->Crit_Chance;
 
 	if (!pFirer)
 		return critChance;
+
+	if (critChance.size() == 0)
+		critChance.push_back(0.0);
 
 	auto const pExt = TechnoExt::ExtMap.Find(pFirer);
 	double extraChance = 0.0;
@@ -456,10 +482,19 @@ double WarheadTypeExt::ExtData::GetCritChance(TechnoClass* pFirer) const
 		if (pType->Crit_DisallowWarheads.size() > 0 && pType->Crit_DisallowWarheads.Contains(this->OwnerObject()))
 			continue;
 
-		critChance = critChance * Math::max(pType->Crit_Multiplier, 0);
+		for (auto& chance : critChance)
+		{
+			chance = chance * Math::max(pType->Crit_Multiplier, 0);
+		}
+
 		extraChance += pType->Crit_ExtraChance;
 	}
 
-	return critChance + extraChance;
+	for (auto& chance : critChance)
+	{
+		chance += extraChance;
+	}
+
+	return critChance;
 }